+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+-Command Queue 0.1.0 (2018-10-03)
+--------------------------------
+- Add `setPaused(boolean)` to temporarily freeze the queue even if receiver is available.
+
 -Command Queue 0.0.3 (2018-09-16)
 --------------------------------
 - Fix possible edge case of a receiver setting a different receiver in the `receiveCommand` block.